WalkMart/requirements.txt
Django==2.0
django-cors-headers==3.2.1
djangorestframework==3.11.0
pytz==2019.3
